WIP Expo 52
Waiting on an update from kysely-expo. It should be as simple as
replacing expo-sqlite/next with expo-sqlite, but I get a lot of
"API call with unopened database connection pointer" logspam with that
change.

Also, not sure why I had to make all these changes to get expo-router
working again. Even with those, the icons in the bottom bar don't load
as expected.